requirements.txt
Flask==2.0.3
mysql-connector-python
jsonpickle
flask_mysqldb
werkzeug